cpu-boost: Boost continuously during input events
At the current moment, the cpu-boost driver only boosts for 40 ms, and
it waits for a total of 150 ms before boosting again. The downside of
this boosting method is that there is an entire 110 ms where the cpus
are not being boosted during input, resulting in potential for stutter.

Make the MIN_INPUT_INTERVAL macro depend on the input boost duration
rather than a fixed value.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>